Add react-redux-router, remove pace.
