+++ package.json
@@ -23,12 +23,15 @@
     "@turf/distance": "^4.5.2",
     "blueimp-md5": "^2.7.0",
     "cheap-ruler": "^2.5.0",
+    "history": "^4.6.3",
     "lodash": "^4.17.4",
     "mapbox-gl": "^0.38.0",
     "prop-types": "^15.5.10",
     "react": "^15.5.0",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.3",
+    "react-router": "^4.1.1",
+    "react-router-redux": "^5.0.0-alpha.6",
     "recharts": "^1.0.0-alpha.0",
     "redux": "^3.6.0",
     "wikidata-sdk": "^4.2.4",
