+++ public/index.html
@@ -8,13 +8,6 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
     <link href="https://api.mapbox.com/mapbox-assembly/v0.12.0/assembly.min.css" rel="stylesheet">
     <script async defer src="https://api.mapbox.com/mapbox-assembly/v0.12.0/assembly.js"></script>
-    <script>
-      window.paceOptions = {
-        ajax: false
-      }
-    </script>
-    <script src="%PUBLIC_URL%/pace.min.js"></script>
-    <link href="%PUBLIC_URL%/pace.css" rel="stylesheet" />
     <title>Mapbox Maps</title>
   </head>
   <body>
