+++ src/components/TrafficSwitch.js
@@ -49,7 +49,7 @@ TrafficSwitch.propTypes = {
 
 const mapStateToProps = (state) => {
   return {
-    mapStyle: state.mapStyle,
+    mapStyle: state.app.mapStyle,
   };
 };
